NerdyPy/requirements.txt
google-api-python-client
lxml==4.3.1
SQLAlchemy==1.3.0
youtube_dl
discord.py[voice]==1.2.3
pydub==0.21.0
aiohttp==3.5.4
